+++ raytrace/cpp/vec3.h
@@ -8,30 +8,30 @@
 class vec3 {
 public:
     vec3() {}
-    vec3(float e0, float e1, float e2) { e[0] = e0; e[1] = e1; e[2] = e2; }
-    inline float x() const { return e[0]; }
-    inline float y() const { return e[1]; }
-    inline float z() const { return e[2]; }
-    inline float r() const { return e[0]; }
-    inline float g() const { return e[1]; }
-    inline float b() const { return e[2]; }
+    vec3(double e0, double e1, double e2) { e[0] = e0; e[1] = e1; e[2] = e2; }
+    inline double x() const { return e[0]; }
+    inline double y() const { return e[1]; }
+    inline double z() const { return e[2]; }
+    inline double r() const { return e[0]; }
+    inline double g() const { return e[1]; }
+    inline double b() const { return e[2]; }
 
-    inline float operator[](int i) const { return e[i]; }
+    inline double operator[](int i) const { return e[i]; }
 
     inline vec3& operator+=(const vec3 &v);
-    inline vec3& operator/=(const float t);
+    inline vec3& operator/=(const double t);
 
-    inline float length() const {
-        return sqrtf(e[0]*e[0] + e[1]*e[1] + e[2]*e[2]);
+    inline double length() const {
+        return sqrt(e[0]*e[0] + e[1]*e[1] + e[2]*e[2]);
     }
 
-    inline float squared_length() const {
+    inline double squared_length() const {
         return e[0]*e[0] + e[1]*e[1] + e[2]*e[2];
     }
 
     friend std::ostream& operator<<(std::ostream& os, const vec3& v);
 
-    float e[3];
+    double e[3];
 };
 
 inline vec3 operator+(const vec3 &v1, const vec3 &v2) {
@@ -42,11 +42,11 @@ inline vec3 operator-(const vec3 &v1, const vec3 &v2) {
     return vec3(v1.e[0] - v2.e[0], v1.e[1] - v2.e[1], v1.e[2] - v2.e[2]);
 }
 
-inline vec3 operator*(float t, const vec3 &v) {
+inline vec3 operator*(double t, const vec3 &v) {
     return vec3(t*v.e[0], t*v.e[1], t*v.e[2]);
 }
 
-inline vec3 operator/(const vec3 &v, float t) {
+inline vec3 operator/(const vec3 &v, double t) {
     return vec3(v.e[0]/t, v.e[1]/t, v.e[2]/t);
 }
 
@@ -57,14 +57,14 @@ inline vec3& vec3::operator+=(const vec3 &v) {
     return *this;
 }
 
-inline vec3& vec3::operator/=(const float t) {
+inline vec3& vec3::operator/=(const double t) {
     e[0] /= t;
     e[1] /= t;
     e[2] /= t;
     return *this;
 }
 
-inline float dot(const vec3 &v1, const vec3 &v2) {
+inline double dot(const vec3 &v1, const vec3 &v2) {
     return v1.e[0] * v2.e[0] + v1.e[1] * v2.e[1] + v1.e[2] * v2.e[2];
 }
